rename: package 이름 수정
memebr -> member
